feat: add minus adjusts
