+++ src/layouts/Integrations/sections/Hero/components/Content/index.tsx
