+++ src/layouts/Integrations/sections/Hero/components/Content/styles.ts
