+++ src/layouts/Integrations/sections/Pipelines/components/Table/index.tsx
