+++ src/layouts/Integrations/sections/Pipelines/components/Table/style.ts
