+++ src/layouts/Integrations/sections/TheAutomation/components/Card/index.tsx
