+++ src/layouts/Integrations/sections/TheAutomation/components/Card/style.ts
