+++ src/layouts/Integrations/sections/TheAutomation/components/Cards/index.tsx
