+++ src/layouts/Integrations/sections/TheAutomation/components/Cards/style.ts
