+++ src/layouts/Integrations/sections/TheAutomation/index.tsx
