+++ src/layouts/Integrations/sections/TheAutomation/style.ts
