+++ src/pages/integrations.tsx
@@ -1,27 +1,28 @@
-import React from 'react'
-import { Connectors } from "../components/Connectors"
-import Layout from "../components/layout"
-import Seo from '../components/seo'
+import React from 'react';
 
-const description = "Capture from clouds, databases, and SaaS apps using real-time ETL, create real-time transformations and materialized views, for a consistent, exact copy of your data powering both analytics and operations."
+import Header from '../layouts/Integrations/components/Head';
+import Wrapper from '../layouts/Integrations/components/Wrapper';
+
+import ChangeData from '../layouts/Integrations/sections/ChangeData';
+import Hero from '../layouts/Integrations/sections/Hero';
+import Pipelines from '../layouts/Integrations/sections/Pipelines';
+import RealTime from '../layouts/Integrations/sections/RealTime';
+import TakeATour from '../layouts/Integrations/sections/TakeATour';
+import TheAutomation from '../layouts/Integrations/sections/TheAutomation';
 
 const Integrations = () => {
-    return (
-        <Layout headerTheme={"light"}>
-            <Connectors
-                title={"Integrations"}
-                description={description}
-                showAllConnectors={true}
-            />
-        </Layout>
-    )
-}
+  return (
+    <Wrapper>
+      <Hero />
+      <ChangeData />
+      <Pipelines />
+      <RealTime />
+      <TakeATour />
+      <TheAutomation />
+    </Wrapper>
+  );
+};
 
-export const Head = () => {
-    return <Seo
-        title={"Integrations"}
-        description={description}
-    />
-}
+export const Head = Header;
 
-export default Integrations
+export default Integrations;
